Add data attributes for toggle and dismiss functionality.
